An integer literal nothing pins down is an Int
It was a BigInt: the type that cannot overflow, chosen so that a program
that never says what its numbers are gets the right answer. The cost was
that every such program ran on heap-allocated numbers and primitive calls
where a machine word and a typed instruction would have done. A program of
the shape a person writes first -- `fun fib n = …`, `def work = [30, 30]`
-- ran in 6.3s and now runs in 0.09s. The cross-language suite is unmoved,
because it annotates everything; the `benches` package, which does not,
rose 20-80x, which also accounts for a "regression" in its header that
could not be bisected: it was measuring the default.

What is given up is that `2 ^ 64` is `0` unless the program says `BigInt`,
with `toBigInt` or an annotation. Std.Num, the tutorial and the tour now
say so where they used to say the opposite, and every test that encoded
the old default checks both: that an unpinned literal wraps, and that a
BigInt asked for is exact.

+++ buildtools/meadow/tests/dap.rs
@@ -89,10 +89,10 @@ fn a_breakpoint_stops_with_the_caller_underneath() {
     assert_eq!(st[1].1, 10, "it resumes at the call: {st:?}");
 
     // `double` is generic over its number type, so `n`'s own type is a
-    // variable; the value says what it is -- a `BigInt`, since `21` in `main`
+    // variable; the value says what it is -- an `Int`, since `21` in `main`
     // is pinned to nothing else.
     let n = local(&mut s, 0, "n");
-    assert_eq!((n.value.as_str(), n.ty.as_deref()), ("21", Some("BigInt")));
+    assert_eq!((n.value.as_str(), n.ty.as_deref()), ("21", Some("Int")));
 }
 
 #[test]
